server/requirements.txt
fastapi
python-multipart
Pillow
imagehash
uvicorn